requirements.txt
# requirements.txt
connexion[swagger-ui]<=2.14.2
Flask==2.2.2
Werkzeug==2.2.2
Flask-SQLAlchemy
psycopg2-binary
PyJWT
swagger-ui-bundle>=0.0.2
python_dateutil>=2.6.0
setuptools>=21.0.0